sync-github-releases: split plan execution into apply-release-plan.ts
sync.ts carried two abstractions: orchestrating the run (pick packages, load
each package's notes, fetch + plan) and executing a plan against GitHub
(applyReleasePlan + resolveCreateTarget). The latter mirrors release-plan.ts —
one computes the plan (pure), the other carries it out (I/O).

Move applyReleasePlan + resolveCreateTarget into apply-release-plan.ts, leaving
sync.ts as pure orchestration (151 -> 100 lines) and release-plan.ts fully pure
and unit-tested. release-plan.ts now exports a named ReleasePlan type — the
plan's shape — in place of applyReleasePlan's prior ReturnType<typeof
getReleasePlan>.
